fix presentation of transactions fields

+++ database/migrations/2023_09_03_113817_create_transactions_table.php
@@ -19,8 +19,8 @@ return new class extends Migration {
             $table->string('iban_other_party');
             $table->string('payment_type');
             $table->string('purpose');
-            $table->integer('value');
-            $table->integer('balance_after');
+            $table->float('value');
+            $table->float('balance_after');
             $table->timestamps();
         });
     }
